+++ printf.h
@@ -6,7 +6,7 @@
 /*   By: deannapiedra <deannapiedra@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 17:16:44 by dpiedra           #+#    #+#             */
-/*   Updated: 2020/02/07 13:30:07 by deannapiedr      ###   ########.fr       */
+/*   Updated: 2020/02/07 14:33:40 by deannapiedr      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,11 +46,11 @@ int				ft_str_precision(t_flag flags, char *str, int count);
 int				ft_neg_precision(t_flag flags, char *str, int count);
 void			ft_putstr_fd(char *s, int fd);
 int				ft_signed_conv(va_list args, t_flag flags);
-int				ft_find_count(int decimal);
-int				ft_signed_flags(t_flag flags, int decimal, int sign, int count);
+int				ft_find_count(long decimal);
+int				ft_signed_flags(t_flag flags, long int decimal, int sign, int count);
 int				ft_signed_precision(t_flag flags, int sign, int count);
 int				ft_signed_width(t_flag flags, int sign, int count);
-void			ft_putnbr_fd(int n, int fd);
+void			ft_putnbr_fd(long int n, int fd);
 int				ft_unsigned_conv(va_list args, t_flag flags);
 int				ft_find_ucount(unsigned int nbr);
 int				ft_unsigned_width(t_flag flags, int count);
